+++ primitives/aesgcm/aesgcm.go
@@ -41,39 +41,46 @@ func AESDecrypt(ciphertext, key []byte) ([]byte, error) {
 }
 
 // EncryptGCM encrypts plaintext using AES-GCM with the provided key and additional data
-func EncryptGCM(plaintext, nonce, key, additionalData []byte) (ciphertext, tag []byte, err error) {
+func AESGCMEncrypt(plaintext,
+	key,
+	initializationVector,
+	additionalAuthenticatedData []byte,
+) (ciphertext, authenticationTag []byte, err error) {
 	block, err := aes.NewCipher(key)
 	if err != nil {
 		return nil, nil, err
 	}
 
-	// Automatically adjust GCM to the nonce size
-	gcm, err := cipher.NewGCMWithNonceSize(block, len(nonce))
+	gcm, err := cipher.NewGCMWithNonceSize(block, len(initializationVector))
 	if err != nil {
 		return nil, nil, err
 	}
 
-	ciphertext = gcm.Seal(nil, nonce, plaintext, additionalData)
-	tag = ciphertext[len(ciphertext)-gcm.Overhead():]
+	ciphertext = gcm.Seal(nil, initializationVector, plaintext, additionalAuthenticatedData)
+	authenticationTag = ciphertext[len(ciphertext)-gcm.Overhead():]
 
-	return ciphertext[:len(ciphertext)-gcm.Overhead()], tag, nil
+	return ciphertext[:len(ciphertext)-gcm.Overhead()], authenticationTag, nil
 }
 
 // DecryptGCM decrypts ciphertext using AES-GCM with the provided key, nonce, additional data, and tag
-func DecryptGCM(ciphertext, key, nonce, additionalData, tag []byte) (plaintext []byte, err error) {
+func AESGCMDecrypt(ciphertext,
+	key,
+	initializationVector,
+	additionalAuthenticatedData,
+	authenticationTag []byte,
+) (plaintext []byte, err error) {
 	block, err := aes.NewCipher(key)
 	if err != nil {
 		return nil, err
 	}
 
-	// Automatically adjust GCM to the nonce size
-	gcm, err := cipher.NewGCMWithNonceSize(block, len(nonce))
+	gcm, err := cipher.NewGCMWithNonceSize(block, len(initializationVector))
 	if err != nil {
 		return nil, err
 	}
 
-	ciphertextWithTag := append(ciphertext, tag...)
-	plaintext, err = gcm.Open(nil, nonce, ciphertextWithTag, additionalData)
+	ciphertextWithTag := append(ciphertext, authenticationTag...)
+	plaintext, err = gcm.Open(nil, initializationVector, ciphertextWithTag, additionalAuthenticatedData)
 	if err != nil {
 		return nil, fmt.Errorf("decryption failed: %w", err)
 	}