+++ primitives/ec/symmetric.go
@@ -19,7 +19,7 @@ func (s *SymmetricKey) Encrypt(message []byte) (ciphertext []byte, err error) {
 	if err != nil {
 		return nil, err
 	}
-	ciphertext, tag, err := aesgcm.EncryptGCM(message, iv, s.ToBytes(), []byte{})
+	ciphertext, tag, err := aesgcm.AESGCMEncrypt(message, s.ToBytes(), iv, []byte{})
 	if err != nil {
 		return nil, err
 	}
@@ -37,7 +37,7 @@ func (s *SymmetricKey) Decrypt(message []byte) (plaintext []byte, err error) {
 	iv := message[:32]
 	ciphertext := message[32 : len(message)-16]
 	tag := message[len(message)-16:]
-	plaintext, err = aesgcm.DecryptGCM(ciphertext, s.ToBytes(), iv, []byte{}, tag)
+	plaintext, err = aesgcm.AESGCMDecrypt(ciphertext, s.ToBytes(), iv, []byte{}, tag)
 	if err != nil {
 		return nil, err
 	}